+++ src/main/java/com/journey/web/repository/member/MemberRepository.java
@@ -1,4 +1,4 @@
-package com.journey.web.repository;
+package com.journey.web.repository.member;
 
 import com.journey.web.domain.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
