+++ src/main/java/com/journey/web/service/MemberService.java
@@ -6,8 +6,8 @@ import com.journey.web.domain.member.MemberStatus;
 import com.journey.web.dto.member.MemberJoinResponseDto;
 import com.journey.web.dto.member.MemberUpdateDto;
 import com.journey.web.exception.CustomException;
-import com.journey.web.repository.MemberItemRepository;
-import com.journey.web.repository.MemberRepository;
+import com.journey.web.repository.member.MemberItemRepository;
+import com.journey.web.repository.member.MemberRepository;
 import com.journey.web.util.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
